requirements.txt
psycopg2==2.9.10
drf-spectacular==0.27.2
djangorestframework-simplejwt==5.3.1
celery==5.3.6
redis==5.0.3
channels-redis==4.2.0
sentry-sdk==2.3.1
locust==2.24.1
opencv-python-headless==4.11.0.86
djangorestframework==3.15.2
channels==4.1.0
daphne==4.1.2
python-dotenv==1.2.2
mediapipe
paho-mqtt
torch
ultralytics
Pillow
